src/requirements.txt
nltk==3.5
numpy==1.19.2
pandas==1.1.3
scikit-learn==0.23.2
scipy==1.5.2
spacy==2.3.2
stop-words==2018.7.23
tensorboard==2.3.0
tensorboardX==2.1
torchvision==0.7.0
tqdm==4.50.2
catalyst==20.9.1
matplotlib==3.3.2